Make test generic to openai compatible endpoints
